actually compare commits

+++ app/pages/index.vue
@@ -1,10 +1,13 @@
 <script setup lang="ts">
 onMounted(() => {
     const commit = useRuntimeConfig().public.gitCommit;
-    console.log('Git Commit:', commit);
-    if (commit !== 'dev' && commit.length > 0) {
+    const oldCommit = localStorage.getItem('commit') || '';
+    console.log('Old commit:', oldCommit);
+    console.log('New commit:', commit);
+    if (commit !== oldCommit && commit !== 'dev' && commit.length > 0) {
         console.warn('COMMIT CHANGED. ERASING LOCAL STORAGE');
         localStorage.clear();
+        localStorage.setItem('commit', commit);
     }
 })
 </script>
